@@ -3,15 +3,12 @@
   "description": "A lightweight web application template built upon the Slim framework and other essential components.",
   "license": "Apache-2.0",
   "require": {
-    "php": ">=7.4",
-    "ext-json": "*",
-    "ext-redis": "*",
+    "php": ">=8.0",
     "ext-PDO": "*",
     "ext-fileinfo": "*",
-    "cache/array-adapter": "^1.0",
-    "cache/prefixed-cache": "^1.0",
-    "cache/psr-6-doctrine-bridge": "^3.0",
-    "cache/redis-adapter": "^1.0",
+    "ext-json": "*",
+    "ext-redis": "*",
+    "azuracast/slim-callable-eventdispatcher": "dev-main",
     "composer/ca-bundle": "^1.2",
     "doctrine/data-fixtures": "^1.3",
     "doctrine/dbal": "^2.8",
@@ -28,38 +25,38 @@
     "php-di/slim-bridge": "^3.0",
     "slim/http": "^1",
     "slim/slim": "^4.2",
+    "symfony/cache": "^5.3",
     "symfony/console": "^5",
     "symfony/event-dispatcher": "^5",
+    "symfony/lock": "^5.3",
+    "symfony/rate-limiter": "^5.3",
     "symfony/serializer": "^5",
     "symfony/validator": "^5"
   },
   "require-dev": {
-    "roave/security-advisories": "dev-master",
-    "codeception/codeception": "^4.0",
-    "codeception/module-phpbrowser": "^1.0",
     "mockery/mockery": "^1.0",
+    "overtrue/phplint": "^1.1",
+    "php-parallel-lint/php-parallel-lint": "^1.3",
     "phpstan/phpstan": "^0.12",
     "phpstan/phpstan-doctrine": "^0.12",
     "phpstan/phpstan-strict-rules": "^0.12",
-    "overtrue/phplint": "^1.1",
-    "codeception/module-asserts": "^1.1"
+    "roave/security-advisories": "dev-latest"
   },
   "replace": {
     "symfony/polyfill-php71": "1.99",
     "symfony/polyfill-php72": "1.99",
     "symfony/polyfill-php73": "1.99",
-    "symfony/polyfill-php74": "1.99"
+    "symfony/polyfill-php74": "1.99",
+    "symfony/polyfill-php80": "1.99"
   },
   "scripts": {
     "test": [
       "@composer install",
       "@phplint",
-      "@phpstan",
-      "@codeception"
+      "@phpstan"
     ],
-    "phplint": "phplint",
-    "phpstan": "phpstan analyze",
-    "codeception": "codecept run --no-interaction --fail-fast"
+    "phplint": "parallel-lint . --exclude vendor",
+    "phpstan": "phpstan analyze"
   },
   "authors": [
     {
