@@ -7,7 +7,7 @@ use Symfony\Contracts\EventDispatcher\Event;
 class BuildConsoleCommands extends Event
 {
     /** @var Application */
-    protected $cli;
+    protected Application $cli;
 
     public function __construct(Application $cli)
     {
