@@ -9,7 +9,7 @@ class BuildRoutes extends Event
     /**
      * @var App
      */
-    protected $app;
+    protected App $app;
 
     public function __construct(App $app)
     {
