@@ -3,7 +3,7 @@ namespace App\Http;
 
 use Psr\Http\Message\ResponseInterface;
 
-final class Response extends \Slim\Http\Response
+class Response extends \Slim\Http\Response
 {
     public const CACHE_ONE_MINUTE = 60;
     public const CACHE_ONE_HOUR = 3600;
@@ -15,7 +15,7 @@ final class Response extends \Slim\Http\Response
      * Send headers that expire the content immediately and prevent caching.
      * @return static
      */
-    public function withNoCache()
+    public function withNoCache(): static
     {
         $response = $this->response
             ->withHeader('Pragma', 'no-cache')
@@ -33,7 +33,7 @@ final class Response extends \Slim\Http\Response
      *
      * @return static
      */
-    public function withCacheLifetime(int $seconds = self::CACHE_ONE_MONTH)
+    public function withCacheLifetime(int $seconds = self::CACHE_ONE_MONTH): static
     {
         $response = $this->response
             ->withHeader('Pragma', '')
@@ -51,10 +51,10 @@ final class Response extends \Slim\Http\Response
     public function hasCacheLifetime(): bool
     {
         if ($this->response->hasHeader('Pragma')) {
-            return (false === strpos($this->response->getHeaderLine('Pragma'), 'no-cache'));
+            return (!str_contains($this->response->getHeaderLine('Pragma'), 'no-cache'));
         }
 
-        return (false === strpos($this->response->getHeaderLine('Cache-Control'), 'no-cache'));
+        return (!str_contains($this->response->getHeaderLine('Cache-Control'), 'no-cache'));
     }
 
     /**
@@ -69,7 +69,7 @@ final class Response extends \Slim\Http\Response
      */
     public function withJson($data, ?int $status = null, int $options = 0, int $depth = 512): ResponseInterface
     {
-        $options |= JSON_UNESCAPED_SLASHES;
+        $options &= JSON_UNESCAPED_SLASHES;
 
         return parent::withJson($data, $status, $options, $depth);
     }
@@ -82,7 +82,7 @@ final class Response extends \Slim\Http\Response
      *
      * @return static
      */
-    public function renderFile($file_path, $file_name = null)
+    public function renderFile(string $file_path, $file_name = null): static
     {
         set_time_limit(600);
 
@@ -113,7 +113,7 @@ final class Response extends \Slim\Http\Response
      *
      * @return static
      */
-    public function renderStringAsFile($file_data, $content_type, $file_name = null)
+    public function renderStringAsFile(string $file_data, string $content_type, string $file_name = null): static
     {
         $response = $this->response
             ->withHeader('Pragma', 'public')
