@@ -3,72 +3,55 @@
 use App\Console\Command;
 use App\Event;
 use App\Middleware;
-use App\Settings;
 use Slim\Interfaces\ErrorHandlerInterface;
 
-return function (\App\EventDispatcher $dispatcher) {
+return function (Azura\SlimCallableEventDispatcher\CallableEventDispatcherInterface $dispatcher) {
     $dispatcher->addListener(Event\BuildConsoleCommands::class, function (Event\BuildConsoleCommands $event) {
         $console = $event->getConsole();
         $di = $console->getContainer();
 
-        /** @var Settings $settings */
-        $settings = $di->get(Settings::class);
+        $console->command('cache:clear', Command\ClearCacheCommand::class)
+            ->setDescription('Clear all application caches.');
 
-        if ($settings->enableRedis()) {
-            $console->command('cache:clear', Command\ClearCacheCommand::class)
-                ->setDescription('Clear all application caches.');
-        }
+        // Doctrine ORM/DBAL
+        Doctrine\ORM\Tools\Console\ConsoleRunner::addCommands($console);
 
-        if ($settings->enableDatabase()) {
-            // Doctrine ORM/DBAL
-            Doctrine\ORM\Tools\Console\ConsoleRunner::addCommands($console);
+        // Add Doctrine Migrations
+        /** @var Doctrine\ORM\EntityManagerInterface $em */
+        $em = $di->get(Doctrine\ORM\EntityManagerInterface::class);
 
-            // Add Doctrine Migrations
-            /** @var Doctrine\ORM\EntityManagerInterface $em */
-            $em = $di->get(Doctrine\ORM\EntityManagerInterface::class);
+        $helper_set = $console->getHelperSet();
+        $doctrine_helpers = Doctrine\ORM\Tools\Console\ConsoleRunner::createHelperSet($em);
+        $helper_set->set($doctrine_helpers->get('db'), 'db');
+        $helper_set->set($doctrine_helpers->get('em'), 'em');
 
-            $helper_set = $console->getHelperSet();
-            $doctrine_helpers = Doctrine\ORM\Tools\Console\ConsoleRunner::createHelperSet($em);
-            $helper_set->set($doctrine_helpers->get('db'), 'db');
-            $helper_set->set($doctrine_helpers->get('em'), 'em');
+        $migrateConfig = new Doctrine\Migrations\Configuration\Migration\ConfigurationArray([
+            'migrations_paths' => [
+                'App\Entity\Migration' => dirname(__DIR__) . '/src/Entity/Migration',
+            ],
+            'table_storage' => [
+                'table_name' => 'app_migrations',
+            ],
+        ]);
 
-            $migrateConfig = new Doctrine\Migrations\Configuration\Migration\ConfigurationArray([
-                'migrations_paths' => [
-                    'App\Entity\Migration' => $settings[Settings::BASE_DIR] . '/src/Entity/Migration',
-                ],
-                'table_storage' => [
-                    'table_name' => 'app_migrations',
-                ],
-            ]);
+        $migrateFactory = Doctrine\Migrations\DependencyFactory::fromEntityManager(
+            $migrateConfig,
+            new Doctrine\Migrations\Configuration\EntityManager\ExistingEntityManager($em)
+        );
+        Doctrine\Migrations\Tools\Console\ConsoleRunner::addCommands($console, $migrateFactory);
 
-            $migrateFactory = Doctrine\Migrations\DependencyFactory::fromEntityManager(
-                $migrateConfig,
-                new Doctrine\Migrations\Configuration\EntityManager\ExistingEntityManager($em)
-            );
-            Doctrine\Migrations\Tools\Console\ConsoleRunner::addCommands($console, $migrateFactory);
-        }
-
-        if (file_exists(__DIR__ . '/cli.php')) {
-            call_user_func(include(__DIR__ . '/cli.php'), $console);
-        }
+        call_user_func(include(__DIR__ . '/cli.php'), $console);
     });
 
     $dispatcher->addListener(Event\BuildRoutes::class, function (Event\BuildRoutes $event) {
         $app = $event->getApp();
 
-        // Load app-specific route configuration.
-        $container = $app->getContainer();
-
-        /** @var Settings $settings */
-        $settings = $container->get(Settings::class);
-
-        if (file_exists($settings[Settings::CONFIG_DIR] . '/routes.php')) {
-            call_user_func(include($settings[Settings::CONFIG_DIR] . '/routes.php'), $app);
-        }
+        call_user_func(include(__DIR__ . '/routes.php'), $app);
 
         // Request injection middlewares.
+        $app->add(Middleware\EnableView::class);
+
         $app->add(Middleware\InjectRouter::class);
-        $app->add(Middleware\InjectRateLimit::class);
 
         // System middleware for routing and body parsing.
         $app->addBodyParsingMiddleware();
@@ -79,7 +62,8 @@ return function (\App\EventDispatcher $dispatcher) {
         $app->add(new Middleware\ApplyXForwardedProto);
 
         // Error handling, which should always be near the "last" element.
-        $errorMiddleware = $app->addErrorMiddleware(!$settings->isProduction(), true, true);
+        $environment = $app->getContainer()->get(App\Environment::class);
+        $errorMiddleware = $app->addErrorMiddleware(!$environment->isProduction(), true, true);
         $errorMiddleware->setDefaultErrorHandler(ErrorHandlerInterface::class);
 
         // Use PSR-7 compatible sessions.
