@@ -1,11 +1,20 @@
 <?php
 
+use App\Environment;
+use Psr\Container\ContainerInterface;
+
 return [
-    // URL Router helper
-    App\Http\Router::class => function (\Slim\App $app, App\Settings $settings) {
-        $route_parser = $app->getRouteCollector()->getRouteParser();
-        return new App\Http\Router($settings, $route_parser);
+    Slim\Interfaces\RouteCollectorInterface::class => static function (Slim\App $app) {
+        return $app->getRouteCollector();
     },
+
+    Slim\Interfaces\RouteParserInterface::class => static function (
+        Slim\Interfaces\RouteCollectorInterface $routeCollector
+    ) {
+        return $routeCollector->getRouteParser();
+    },
+
+    // URL Router helper
     App\Http\RouterInterface::class => DI\Get(App\Http\Router::class),
 
     // Error handler
@@ -37,20 +46,43 @@ return [
     },
 
     // Cache
-    Psr\Cache\CacheItemPoolInterface::class => function (App\Settings $settings, Psr\Container\ContainerInterface $di) {
-        // Never use the Redis cache for CLI commands, as the CLI commands are where
-        // the Redis cache gets flushed, so this will lead to a race condition that can't
-        // be solved within the application.
-        return $settings->enableRedis() && !$settings->isCli()
-            ? new Cache\Adapter\Redis\RedisCachePool($di->get(Redis::class))
-            : new Cache\Adapter\PHPArray\ArrayCachePool;
-    },
-    Psr\SimpleCache\CacheInterface::class => DI\get(Psr\Cache\CacheItemPoolInterface::class),
+    Redis::class => static function (Environment $environment) {
+        $settings = $environment->getRedisSettings();
 
-    // Configuration management
-    App\Config::class => function (App\Settings $settings) {
-        return new App\Config($settings[App\Settings::CONFIG_DIR]);
+        $redis = new Redis();
+        $redis->connect($settings['host'], $settings['port'], 15);
+        $redis->select($settings['db']);
+
+        return $redis;
     },
+
+    Symfony\Contracts\Cache\CacheInterface::class => static function (
+        Environment $environment,
+        Psr\Log\LoggerInterface $logger,
+        ContainerInterface $di
+    ) {
+        /** @var Symfony\Contracts\Cache\CacheInterface $cacheInterface */
+        if ($environment->isTesting()) {
+            $cacheInterface = new Symfony\Component\Cache\Adapter\ArrayAdapter();
+        } else {
+            $tempDir = $environment->getTempDirectory() . DIRECTORY_SEPARATOR . 'cache';
+            $cacheInterface = new Symfony\Component\Cache\Adapter\FilesystemAdapter(
+                '',
+                0,
+                $tempDir
+            );
+        }
+
+        $cacheInterface->setLogger($logger);
+        return $cacheInterface;
+    },
+
+    Symfony\Component\Cache\Adapter\AdapterInterface::class => DI\get(
+        Symfony\Contracts\Cache\CacheInterface::class
+    ),
+    Psr\Cache\CacheItemPoolInterface::class => DI\get(
+        Symfony\Contracts\Cache\CacheInterface::class
+    ),
 
     // DBAL
     Doctrine\DBAL\Connection::class => function (Doctrine\ORM\EntityManager $em) {
@@ -58,7 +90,10 @@ return [
     },
 
     // Console
-    App\Console\Application::class => function (DI\Container $di, App\EventDispatcher $dispatcher) {
+    App\Console\Application::class => function (
+        DI\Container $di,
+        Azura\SlimCallableEventDispatcher\CallableEventDispatcherInterface $dispatcher
+    ) {
         $console = new App\Console\Application('Command Line Interface', '1.0.0', $di);
 
         // Trigger an event for the core app and all plugins to build their CLI commands.
@@ -69,70 +104,68 @@ return [
     },
 
     // Doctrine cache
-    Doctrine\Common\Cache\Cache::class => function (Psr\Cache\CacheItemPoolInterface $cachePool) {
-        return new Cache\Bridge\Doctrine\DoctrineCacheBridge(new Cache\Prefixed\PrefixedCachePool($cachePool,
-            'doctrine|'));
+    Doctrine\Common\Cache\Cache::class => static function (
+        Environment $environment,
+        Psr\Cache\CacheItemPoolInterface $psr6Cache
+    ) {
+        if ($environment->isCli()) {
+            $psr6Cache = new Symfony\Component\Cache\Adapter\ArrayAdapter();
+        }
+
+        $proxyCache = new Symfony\Component\Cache\Adapter\ProxyAdapter($psr6Cache, 'doctrine.');
+        return Doctrine\Common\Cache\Psr6\DoctrineProvider::wrap($proxyCache);
     },
 
     // Doctrine Entity Manager
-    App\Doctrine\DecoratedEntityManager::class => function (
+    App\Doctrine\DecoratedEntityManager::class => static function (
         Doctrine\Common\Cache\Cache $doctrineCache,
-        Doctrine\Common\Annotations\Reader $reader,
-        App\Settings $settings
+        Environment $environment
     ) {
-        $connectionOptions = [
-            'host' => $_ENV['MYSQL_HOST'] ?? 'mariadb',
-            'port' => $_ENV['MYSQL_PORT'] ?? 3306,
-            'dbname' => $_ENV['MYSQL_DATABASE'],
-            'user' => $_ENV['MYSQL_USER'],
-            'password' => $_ENV['MYSQL_PASSWORD'],
-            'driver' => 'pdo_mysql',
-            'charset' => 'utf8mb4',
-            'defaultTableOptions' => [
+        $connectionOptions = array_merge(
+            $environment->getDatabaseSettings(),
+            [
+                'driver' => 'pdo_mysql',
                 'charset' => 'utf8mb4',
-                'collate' => 'utf8mb4_general_ci',
-            ],
-            'driverOptions' => [
-                // PDO::MYSQL_ATTR_INIT_COMMAND = 1002;
-                1002 => 'SET NAMES utf8mb4 COLLATE utf8mb4_general_ci',
-            ],
-            'platform' => new Doctrine\DBAL\Platforms\MariaDb1027Platform(),
-        ];
-
-        if (!$settings[App\Settings::IS_DOCKER]) {
-            $connectionOptions['host'] = $_ENV['db_host'] ?? 'localhost';
-            $connectionOptions['port'] = $_ENV['db_port'] ?? '3306';
-            $connectionOptions['dbname'] = $_ENV['db_name'] ?? 'azuracast';
-            $connectionOptions['user'] = $_ENV['db_username'] ?? 'azuracast';
-            $connectionOptions['password'] = $_ENV['db_password'];
-        }
+                'defaultTableOptions' => [
+                    'charset' => 'utf8mb4',
+                    'collate' => 'utf8mb4_general_ci',
+                ],
+                'driverOptions' => [
+                    // PDO::MYSQL_ATTR_INIT_COMMAND = 1002;
+                    1002 => 'SET NAMES utf8mb4 COLLATE utf8mb4_general_ci',
+                ],
+                'platform' => new Doctrine\DBAL\Platforms\MariaDb1027Platform(),
+            ]
+        );
 
         try {
             // Fetch and store entity manager.
             $config = Doctrine\ORM\Tools\Setup::createConfiguration(
                 Doctrine\Common\Proxy\AbstractProxyFactory::AUTOGENERATE_FILE_NOT_EXISTS,
-                $settings->getTempDirectory() . '/proxies',
+                $environment->getTempDirectory() . '/proxies',
                 $doctrineCache
             );
 
-            $annotationDriver = new Doctrine\ORM\Mapping\Driver\AnnotationDriver(
-                $reader,
-                [$settings->getBaseDirectory() . '/src/Entity']
+            $mappingClassesPaths = [$environment->getBaseDirectory() . '/src/Entity'];
+            $attributeDriver = new Doctrine\ORM\Mapping\Driver\AttributeDriver(
+                $mappingClassesPaths
             );
-            $config->setMetadataDriverImpl($annotationDriver);
+            $config->setMetadataDriverImpl($attributeDriver);
 
             // Debug mode:
             // $config->setSQLLogger(new Doctrine\DBAL\Logging\EchoSQLLogger);
 
             $eventManager = new Doctrine\Common\EventManager;
 
-            return new App\Doctrine\DecoratedEntityManager(function () use (
-                $connectionOptions,
-                $config,
-                $eventManager
-            ) {
-                return Doctrine\ORM\EntityManager::create($connectionOptions, $config, $eventManager);
-            });
+            return new App\Doctrine\DecoratedEntityManager(
+                function () use (
+                    $connectionOptions,
+                    $config,
+                    $eventManager
+                ) {
+                    return Doctrine\ORM\EntityManager::create($connectionOptions, $config, $eventManager);
+                }
+            );
         } catch (Exception $e) {
             throw new App\Exception\BootstrapException($e->getMessage());
         }
@@ -140,8 +173,10 @@ return [
     Doctrine\ORM\EntityManagerInterface::class => DI\Get(App\Doctrine\DecoratedEntityManager::class),
 
     // Event Dispatcher
-    App\EventDispatcher::class => function (Slim\App $app) {
-        $dispatcher = new App\EventDispatcher($app->getCallableResolver());
+    Azura\SlimCallableEventDispatcher\CallableEventDispatcherInterface::class => static function (
+        Slim\App $app
+    ) {
+        $dispatcher = new Azura\SlimCallableEventDispatcher\SlimCallableEventDispatcher($app->getCallableResolver());
 
         // Register application default events.
         if (file_exists(__DIR__ . '/events.php')) {
@@ -150,73 +185,57 @@ return [
 
         return $dispatcher;
     },
+    Psr\EventDispatcher\EventDispatcherInterface::class => DI\get(
+        Azura\SlimCallableEventDispatcher\CallableEventDispatcherInterface::class
+    ),
 
     // Monolog Logger
-    Monolog\Logger::class => function (App\Settings $settings) {
-        $logger = new Monolog\Logger($settings[App\Settings::APP_NAME] ?? 'app');
-        $logging_level = $settings->isProduction() ? Psr\Log\LogLevel::INFO : Psr\Log\LogLevel::DEBUG;
+    Monolog\Logger::class => static function (Environment $environment) {
+        $logger = new Monolog\Logger($environment->getAppName());
+        $loggingLevel = Psr\Log\LogLevel::INFO;
 
-        if ($settings[App\Settings::IS_DOCKER] || $settings[App\Settings::IS_CLI]) {
-            $log_stderr = new Monolog\Handler\StreamHandler('php://stderr', $logging_level, true);
+        if ($environment->isCli()) {
+            $log_stderr = new Monolog\Handler\StreamHandler('php://stderr', $loggingLevel, true);
             $logger->pushHandler($log_stderr);
         }
 
-        $log_file = new Monolog\Handler\StreamHandler($settings[App\Settings::TEMP_DIR] . '/app.log',
-            $logging_level, true);
+        $log_file = new Monolog\Handler\RotatingFileHandler(
+            $environment->getTempDirectory() . '/app.log',
+            5,
+            $loggingLevel,
+            true
+        );
         $logger->pushHandler($log_file);
 
         return $logger;
     },
     Psr\Log\LoggerInterface::class => DI\get(Monolog\Logger::class),
 
-    // Session save handler middleware
-    Mezzio\Session\SessionPersistenceInterface::class => function (Cache\Adapter\Redis\RedisCachePool $redisPool) {
-        return new Mezzio\Session\Cache\CacheSessionPersistence(
-            new Cache\Prefixed\PrefixedCachePool($redisPool, 'session|'),
-            'app_session',
-            '/',
-            'nocache',
-            43200,
-            time()
-        );
-    },
-
-    // Redis cache
-    Redis::class => function (App\Settings $settings) {
-        $redis_host = $settings[App\Settings::IS_DOCKER] ? 'redis' : 'localhost';
-
-        $redis = new Redis();
-        $redis->connect($redis_host, 6379, 15);
-        $redis->select(1);
-
-        return $redis;
-    },
-
     // Doctrine annotations reader
     Doctrine\Common\Annotations\Reader::class => function (
         Doctrine\Common\Cache\Cache $doctrine_cache,
-        App\Settings $settings
+        Environment $environment
     ) {
         return new Doctrine\Common\Annotations\CachedReader(
             new Doctrine\Common\Annotations\AnnotationReader,
             $doctrine_cache,
-            !$settings->isProduction()
+            !$environment->isProduction()
         );
     },
 
     // Symfony Serializer
-    Symfony\Component\Serializer\Serializer::class => function (
-        Doctrine\Common\Annotations\Reader $annotation_reader,
-        Doctrine\ORM\EntityManager $em
+    Symfony\Component\Serializer\Serializer::class => static function (
+        Doctrine\Common\Annotations\Reader $reader,
+        Doctrine\ORM\EntityManagerInterface $em
     ) {
-        $meta_factory = new Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory(
-            new Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader($annotation_reader)
+        $classMetaFactory = new Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory(
+            new Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader($reader)
         );
 
         $normalizers = [
             new Symfony\Component\Serializer\Normalizer\JsonSerializableNormalizer(),
-            new App\Normalizer\DoctrineEntityNormalizer($em, $annotation_reader, $meta_factory),
-            new Symfony\Component\Serializer\Normalizer\ObjectNormalizer($meta_factory),
+            new App\Normalizer\DoctrineEntityNormalizer($em, $classMetaFactory),
+            new Symfony\Component\Serializer\Normalizer\ObjectNormalizer($classMetaFactory),
         ];
         $encoders = [
             new Symfony\Component\Serializer\Encoder\JsonEncoder,
@@ -236,5 +255,13 @@ return [
         $builder->setConstraintValidatorFactory($cvf);
         $builder->enableAnnotationMapping($annotation_reader);
         return $builder->getValidator();
+    },
+
+    // Symfony Lock adapter
+    Symfony\Component\Lock\PersistingStoreInterface::class => static function (
+        ContainerInterface $di
+    ) {
+        $redis = $di->get(Redis::class);
+        return new Symfony\Component\Lock\Store\RedisStore($redis);
     },
 ];
