@@ -1,8 +1,12 @@
 <?php
+
+declare(strict_types=1);
+
 namespace App\Http;
 
-use App\Exception;
-use App\Settings;
+use App\Entity;
+use App\Environment;
+use App\Traits\RequestAwareTrait;
 use GuzzleHttp\Psr7\Uri;
 use GuzzleHttp\Psr7\UriResolver;
 use InvalidArgumentException;
@@ -14,30 +18,148 @@ use Slim\Routing\RouteContext;
 
 class Router implements RouterInterface
 {
-    protected RouteParserInterface $route_parser;
+    use RequestAwareTrait;
 
-    protected Settings $settings;
+    protected ?UriInterface $baseUrl = null;
 
-    protected ServerRequestInterface $current_request;
+    public function __construct(
+        protected Environment $environment,
+        protected RouteParserInterface $routeParser
+    ) {
+    }
 
-    public function __construct(Settings $settings, RouteParserInterface $route_parser)
+    public function setRequest(?ServerRequestInterface $request): void
     {
-        $this->settings = $settings;
-        $this->route_parser = $route_parser;
+        $this->request = $request;
+        $this->baseUrl = null;
+    }
+
+    public function getBaseUrl(bool $useRequest = true): UriInterface
+    {
+        if (null === $this->baseUrl) {
+            $this->baseUrl = $this->buildBaseUrl();
+        }
+
+        return $this->baseUrl;
+    }
+
+    protected function buildBaseUrl(): UriInterface
+    {
+        $baseUrl = new Uri('');
+        $useHttps = false;
+
+        if ($this->request instanceof ServerRequestInterface) {
+            $currentUri = $this->request->getUri();
+
+            if ('https' === $currentUri->getScheme()) {
+                $useHttps = true;
+            }
+
+            if ($baseUrl->getHost() === '') {
+                $ignoredHosts = ['web', 'nginx', 'localhost'];
+                if (!in_array($currentUri->getHost(), $ignoredHosts, true)) {
+                    $baseUrl = (new Uri())
+                        ->withScheme($currentUri->getScheme())
+                        ->withHost($currentUri->getHost())
+                        ->withPort($currentUri->getPort());
+                }
+            }
+        }
+
+        if ($useHttps && $baseUrl->getScheme() !== '') {
+            $baseUrl = $baseUrl->withScheme('https');
+        }
+
+        // Avoid double-trailing slashes in various URLs
+        if ('/' === $baseUrl->getPath()) {
+            $baseUrl = $baseUrl->withPath('');
+        }
+
+        // Filter the base URL so it doesn't say http://site:80 or https://site:443
+        if (Uri::isDefaultPort($baseUrl)) {
+            return $baseUrl->withPort(null);
+        }
+
+        return $baseUrl;
     }
 
     /**
-     * Compose a URL, returning an absolute URL (including base URL) if the current settings or this function's parameters
-     * indicate an absolute URL is necessary
+     * @inheritDoc
+     */
+    public function fromHereWithQuery(
+        ?string $route_name = null,
+        array $route_params = [],
+        array $query_params = [],
+        bool $absolute = false
+    ): UriInterface {
+        if ($this->request instanceof ServerRequestInterface) {
+            $query_params = array_merge($this->request->getQueryParams(), $query_params);
+        }
+
+        return $this->fromHere($route_name, $route_params, $query_params, $absolute);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function fromHere(
+        ?string $route_name = null,
+        array $route_params = [],
+        array $query_params = [],
+        bool $absolute = false
+    ): UriInterface {
+        if ($this->request instanceof ServerRequestInterface) {
+            $route = RouteContext::fromRequest($this->request)->getRoute();
+        } else {
+            $route = null;
+        }
+
+        if (null === $route_name && $route instanceof RouteInterface) {
+            $route_name = $route->getName();
+        }
+
+        if (null === $route_name) {
+            throw new InvalidArgumentException(
+                'Cannot specify a null route name if no existing route is configured.'
+            );
+        }
+
+        if ($route instanceof RouteInterface) {
+            $route_params = array_merge($route->getArguments(), $route_params);
+        }
+
+        return $this->named($route_name, $route_params, $query_params, $absolute);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function named(
+        string $route_name,
+        array $route_params = [],
+        array $query_params = [],
+        bool $absolute = false
+    ): UriInterface {
+        return self::resolveUri(
+            $this->getBaseUrl(),
+            $this->routeParser->relativeUrlFor($route_name, $route_params, $query_params),
+            $absolute
+        );
+    }
+
+    /**
+     * Compose a URL, returning an absolute URL (including base URL) if the current settings or
+     * this function's parameters indicate an absolute URL is necessary
      *
      * @param UriInterface $base
-     * @param UriInterface|string $rel
+     * @param string|UriInterface $rel
      * @param bool $absolute
-     *
-     * @return UriInterface
      */
-    public static function resolveUri(UriInterface $base, $rel, bool $absolute = false): UriInterface
-    {
+    public static function resolveUri(
+        UriInterface $base,
+        UriInterface|string $rel,
+        bool $absolute = false
+    ): UriInterface {
         if (!$rel instanceof UriInterface) {
             $rel = new Uri($rel);
         }
@@ -56,134 +178,5 @@ class Router implements RouterInterface
         }
 
         return UriResolver::resolve($base, $rel);
-    }
-
-    /**
-     * @return ServerRequestInterface
-     */
-    public function getCurrentRequest(): ServerRequestInterface
-    {
-        return $this->current_request;
-    }
-
-    /**
-     * @param ServerRequestInterface $current_request
-     */
-    public function setCurrentRequest(ServerRequestInterface $current_request): void
-    {
-        $this->current_request = $current_request;
-    }
-
-    /**
-     * Same as $this->fromHere(), but merging the current GET query parameters into the request as well.
-     *
-     * @param null $route_name
-     * @param array $route_params
-     * @param array $query_params
-     * @param bool $absolute
-     *
-     * @return string
-     */
-    public function fromHereWithQuery(
-        $route_name = null,
-        array $route_params = [],
-        array $query_params = [],
-        $absolute = false
-    ): string {
-        if ($this->current_request instanceof ServerRequestInterface) {
-            $query_params = array_merge($this->current_request->getQueryParams(), $query_params);
-        }
-
-        return $this->fromHere($route_name, $route_params, $query_params, $absolute);
-    }
-
-    /**
-     * Return a named route based on the current page and its route arguments.
-     *
-     * @param null $route_name
-     * @param array $route_params
-     * @param array $query_params
-     * @param bool $absolute
-     *
-     * @return string
-     */
-    public function fromHere(
-        $route_name = null,
-        array $route_params = [],
-        array $query_params = [],
-        $absolute = false
-    ): string {
-        if ($this->current_request instanceof ServerRequestInterface) {
-            $routeContext = RouteContext::fromRequest($this->current_request);
-            $route = $routeContext->getRoute();
-        } else {
-            $route = null;
-        }
-
-        if ($route_name === null) {
-            if ($route instanceof RouteInterface) {
-                $route_name = $route->getName();
-            } else {
-                throw new InvalidArgumentException('Cannot specify a null route name if no existing route is configured.');
-            }
-        }
-
-        if ($route instanceof RouteInterface) {
-            $route_params = array_merge($route->getArguments(), $route_params);
-        }
-
-        return $this->named($route_name, $route_params, $query_params, $absolute);
-    }
-
-    /**
-     * Simpler format for calling "named" routes with parameters.
-     *
-     * @param string $route_name
-     * @param array $route_params
-     * @param array $query_params
-     * @param boolean $absolute Whether to include the full URL.
-     *
-     * @return UriInterface
-     */
-    public function named($route_name, $route_params = [], array $query_params = [], $absolute = false): UriInterface
-    {
-        return self::resolveUri($this->getBaseUrl(),
-            $this->route_parser->relativeUrlFor($route_name, $route_params, $query_params), $absolute);
-    }
-
-    /**
-     * Dynamically calculate the base URL the first time it's called, if it is at all in the request.
-     *
-     * @param bool $use_request Use the current request for the base URI, if available.
-     *
-     * @return UriInterface
-     */
-    public function getBaseUrl(bool $use_request = true): UriInterface
-    {
-        static $base_url;
-
-        // Check the settings for a hard-coded base URI.
-        if (!$base_url && !empty($this->settings[Settings::BASE_URL])) {
-            $base_url = new Uri($this->settings[Settings::BASE_URL]);
-        }
-
-        // Use the current request's URI if applicable.
-        if ($use_request && $this->current_request instanceof ServerRequestInterface) {
-            $current_uri = $this->current_request->getUri();
-
-            $ignored_hosts = ['nginx', 'localhost'];
-            if (!in_array($current_uri->getHost(), $ignored_hosts)) {
-                return (new Uri())
-                    ->withScheme($current_uri->getScheme())
-                    ->withHost($current_uri->getHost())
-                    ->withPort($current_uri->getPort());
-            }
-        }
-
-        if (!$base_url instanceof Uri) {
-            throw new Exception('Base URL could not be determined.');
-        }
-
-        return $base_url;
     }
 }
