@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace App\Http;
 
 use Psr\Http\Message\ServerRequestInterface;
@@ -6,15 +9,11 @@ use Psr\Http\Message\UriInterface;
 
 interface RouterInterface
 {
-    /**
-     * @param ServerRequestInterface $current_request
-     */
-    public function setCurrentRequest(ServerRequestInterface $current_request): void;
+    public function setRequest(?ServerRequestInterface $request): void;
 
-    /**
-     * @return ServerRequestInterface
-     */
-    public function getCurrentRequest(): ServerRequestInterface;
+    public function withRequest(?ServerRequestInterface $request): self;
+
+    public function getBaseUrl(): UriInterface;
 
     /**
      * Simpler format for calling "named" routes with parameters.
@@ -23,51 +22,41 @@ interface RouterInterface
      * @param array $route_params
      * @param array $query_params
      * @param boolean $absolute Whether to include the full URL.
-     *
-     * @return UriInterface
      */
-    public function named($route_name, $route_params = [], array $query_params = [], $absolute = false): UriInterface;
-
-    /**
-     * Dynamically calculate the base URL the first time it's called, if it is at all in the request.
-     *
-     * @param bool $use_request Use the current request for the base URI, if available.
-     *
-     * @return UriInterface
-     */
-    public function getBaseUrl(bool $use_request = true): UriInterface;
+    public function named(
+        string $route_name,
+        array $route_params = [],
+        array $query_params = [],
+        bool $absolute = false
+    ): UriInterface;
 
     /**
      * Return a named route based on the current page and its route arguments.
      *
-     * @param null $route_name
+     * @param string|null $route_name
      * @param array $route_params
      * @param array $query_params
      * @param bool $absolute
-     *
-     * @return string
      */
     public function fromHere(
-        $route_name = null,
+        ?string $route_name = null,
         array $route_params = [],
         array $query_params = [],
-        $absolute = false
-    ): string;
+        bool $absolute = false
+    ): UriInterface;
 
     /**
      * Same as $this->fromHere(), but merging the current GET query parameters into the request as well.
      *
-     * @param null $route_name
+     * @param string|null $route_name
      * @param array $route_params
      * @param array $query_params
      * @param bool $absolute
-     *
-     * @return string
      */
     public function fromHereWithQuery(
-        $route_name = null,
+        ?string $route_name = null,
         array $route_params = [],
         array $query_params = [],
-        $absolute = false
-    ): string;
+        bool $absolute = false
+    ): UriInterface;
 }
