@@ -1,45 +1,64 @@
 <?php
+
+declare(strict_types=1);
+
 namespace App\Middleware;
 
+use App\Environment;
 use App\Http\ServerRequest;
 use App\Session\Csrf;
 use App\Session\Flash;
-use App\Settings;
+use Mezzio\Session\Cache\CacheSessionPersistence;
 use Mezzio\Session\LazySession;
-use Mezzio\Session\SessionPersistenceInterface;
+use Psr\Cache\CacheItemPoolInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
+use Symfony\Component\Cache\Adapter\ArrayAdapter;
+use Symfony\Component\Cache\Adapter\ProxyAdapter;
 
 /**
  * Inject the session object into the request.
  */
 class InjectSession implements MiddlewareInterface
 {
-    protected SessionPersistenceInterface $sessionPersistence;
-
-    protected Settings $settings;
+    protected CacheItemPoolInterface $cachePool;
 
     public function __construct(
-        SessionPersistenceInterface $sessionPersistence,
-        Settings $settings
+        CacheItemPoolInterface $cachePool,
+        protected Environment $environment
     ) {
-        $this->sessionPersistence = $sessionPersistence;
-        $this->settings = $settings;
+        if ($environment->isCli()) {
+            $cachePool = new ArrayAdapter();
+        }
+
+        $this->cachePool = new ProxyAdapter($cachePool, 'session.');
     }
 
     /**
      * @param ServerRequestInterface $request
      * @param RequestHandlerInterface $handler
-     *
-     * @return ResponseInterface
      */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
-        $session = new LazySession($this->sessionPersistence, $request);
+        $isHttpsUrl = ('https' === $request->getUri()->getScheme());
 
-        $csrf = new Csrf($session, $this->settings);
+        $sessionPersistence = new CacheSessionPersistence(
+            cache: $this->cachePool,
+            cookieName: 'app_session',
+            cookiePath: '/',
+            cacheLimiter: 'nocache',
+            cacheExpire: 43200,
+            lastModified: time(),
+            persistent: true,
+            cookieSecure: $isHttpsUrl,
+            cookieHttpOnly: true
+        );
+
+        $session = new LazySession($sessionPersistence, $request);
+
+        $csrf = new Csrf($session, $this->environment);
         Csrf::setInstance($csrf);
 
         $flash = new Flash($session);
@@ -50,6 +69,6 @@ class InjectSession implements MiddlewareInterface
             ->withAttribute(ServerRequest::ATTR_SESSION_FLASH, $flash);
 
         $response = $handler->handle($request);
-        return $this->sessionPersistence->persistSession($session, $response);
+        return $sessionPersistence->persistSession($session, $response);
     }
 }
