@@ -1,86 +1,51 @@
 <?php
+
+declare(strict_types=1);
+
 namespace App\Normalizer;
 
 use App\Exception\NoGetterAvailableException;
-use App\Normalizer\Annotation\DeepNormalize;
-use DateTime;
-use Doctrine\Common\Annotations\Reader;
+use App\Normalizer\Attributes\DeepNormalize;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\Inflector\Inflector;
 use Doctrine\Inflector\InflectorFactory;
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\ORM\Mapping\Driver\AnnotationDriver;
 use Doctrine\ORM\Proxy\Proxy;
-use InvalidArgumentException;
 use ProxyManager\Proxy\GhostObjectInterface;
 use ReflectionClass;
 use ReflectionMethod;
 use ReflectionNamedType;
+use ReflectionParameter;
 use ReflectionProperty;
-use Symfony\Component\Serializer\Mapping\AttributeMetadataInterface;
-use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
+use RuntimeException;
 use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactoryInterface;
-use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
-use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
-use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
-use Symfony\Component\Serializer\Normalizer\ObjectToPopulateTrait;
-use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
+use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
 use function is_array;
 
-class DoctrineEntityNormalizer extends AbstractNormalizer
+class DoctrineEntityNormalizer extends AbstractNormalizer implements NormalizerAwareInterface
 {
-    use ObjectToPopulateTrait;
+    use NormalizerAwareTrait;
 
     public const NORMALIZE_TO_IDENTIFIERS = 'form_mode';
 
     public const CLASS_METADATA = 'class_metadata';
     public const ASSOCIATION_MAPPINGS = 'association_mappings';
 
-    /** @var SerializerInterface|NormalizerInterface|DenormalizerInterface */
-    protected $serializer;
-
-    protected EntityManagerInterface $em;
-
-    protected Reader $annotationReader;
-
     protected Inflector $inflector;
 
     public function __construct(
-        EntityManagerInterface $em,
-        Reader $annotationReader = null,
+        protected EntityManagerInterface $em,
         ClassMetadataFactoryInterface $classMetadataFactory = null,
         NameConverterInterface $nameConverter = null,
         array $defaultContext = []
     ) {
-        /** @var AnnotationDriver $metadata_driver */
-        $metadata_driver = $em->getConfiguration()->getMetadataDriverImpl();
-
-        $annotationReader = $annotationReader ?? $metadata_driver->getReader();
-        $classMetadataFactory = $classMetadataFactory ?? new ClassMetadataFactory(
-                new AnnotationLoader($annotationReader)
-            );
-
         $defaultContext[self::ALLOW_EXTRA_ATTRIBUTES] = false;
-
         parent::__construct($classMetadataFactory, $nameConverter, $defaultContext);
 
-        $this->em = $em;
-        $this->annotationReader = $annotationReader;
         $this->inflector = InflectorFactory::create()->build();
-    }
-
-    /**
-     * @param DenormalizerInterface|NormalizerInterface|SerializerInterface $serializer
-     */
-    public function setSerializer($serializer): void
-    {
-        if (!$serializer instanceof DenormalizerInterface || !$serializer instanceof NormalizerInterface) {
-            throw new InvalidArgumentException('Expected a serializer that also implements DenormalizerInterface and NormalizerInterface.');
-        }
-
-        $this->serializer = $serializer;
     }
 
     /**
@@ -90,9 +55,8 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
      * @param string|null $format
      * @param array $context
      *
-     * @return mixed
      */
-    public function normalize($object, string $format = null, array $context = [])
+    public function normalize($object, string $format = null, array $context = []): mixed
     {
         if ($this->isCircularReference($object, $context)) {
             return $this->handleCircularReference($object, $format, $context);
@@ -111,14 +75,16 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
                     $value = $this->getAttributeValue($object, $attribute, $format, $context);
 
                     /** @var callable|null $callback */
-                    $callback = $context[self::CALLBACKS][$attribute] ?? $this->defaultContext[self::CALLBACKS][$attribute] ?? null;
+                    $callback = $context[self::CALLBACKS][$attribute]
+                        ?? $this->defaultContext[self::CALLBACKS][$attribute]
+                        ?? null;
+
                     if ($callback) {
                         $value = $callback($value, $object, $attribute, $format, $context);
                     }
 
                     $return_arr[$attribute] = $value;
-                } catch (NoGetterAvailableException $e) {
-                    continue;
+                } catch (NoGetterAvailableException) {
                 }
             }
         }
@@ -130,19 +96,17 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
      * Replicates the "fromArray" functionality previously present in Doctrine 1.
      *
      * @param mixed $data
-     * @param string $class
+     * @param class-string $type
      * @param string|null $format
      * @param array $context
-     *
-     * @return object
      */
-    public function denormalize($data, $class, string $format = null, array $context = [])
+    public function denormalize($data, string $type, string $format = null, array $context = []): object
     {
-        $object = $this->instantiateObject($data, $class, $context, new ReflectionClass($class), false, $format);
+        $object = $this->instantiateObject($data, $type, $context, new ReflectionClass($type), false, $format);
 
-        $class = get_class($object);
+        $type = get_class($object);
 
-        $context[self::CLASS_METADATA] = $this->em->getMetadataFactory()->getMetadataFor($class);
+        $context[self::CLASS_METADATA] = $this->em->getMetadataFactory()->getMetadataFor($type);
         $context[self::ASSOCIATION_MAPPINGS] = [];
 
         if ($context[self::CLASS_METADATA]->associationMappings) {
@@ -172,7 +136,10 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
 
         foreach ((array)$data as $attribute => $value) {
             /** @var callable|null $callback */
-            $callback = $context[self::CALLBACKS][$attribute] ?? $this->defaultContext[self::CALLBACKS][$attribute] ?? null;
+            $callback = $context[self::CALLBACKS][$attribute]
+                ?? $this->defaultContext[self::CALLBACKS][$attribute]
+                ?? null;
+
             if ($callback) {
                 $value = $callback($value, $object, $attribute, $format, $context);
             }
@@ -186,7 +153,7 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
     /**
      * @inheritdoc
      */
-    public function supportsNormalization($data, string $format = null)
+    public function supportsNormalization($data, string $format = null): bool
     {
         return $this->isEntity($data);
     }
@@ -194,24 +161,30 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
     /**
      * @inheritdoc
      */
-    public function supportsDenormalization($data, $type, string $format = null)
+    public function supportsDenormalization($data, $type, string $format = null): bool
     {
         return $this->isEntity($type);
     }
 
     /**
-     * @param object|string $classOrObject
+     * @param object|class-string $classOrObject
      * @param array $context
      * @param bool $attributesAsString
      *
-     * @return bool|string[]|AttributeMetadataInterface[]
      */
-    protected function getAllowedAttributes($classOrObject, array $context, $attributesAsString = false)
-    {
-        $meta = $this->classMetadataFactory->getMetadataFor($classOrObject)->getAttributesMetadata();
+    protected function getAllowedAttributes(
+        $classOrObject,
+        array $context,
+        bool $attributesAsString = false
+    ): array|false {
+        $meta = $this->classMetadataFactory?->getMetadataFor($classOrObject)?->getAttributesMetadata();
+        if (null === $meta) {
+            throw new RuntimeException('Class metadata factory not specified.');
+        }
 
-        $reflect = new ReflectionClass($classOrObject);
-        $props_raw = $reflect->getProperties(ReflectionProperty::IS_PUBLIC | ReflectionProperty::IS_PROTECTED);
+        $props_raw = (new ReflectionClass($classOrObject))->getProperties(
+            ReflectionProperty::IS_PUBLIC | ReflectionProperty::IS_PROTECTED
+        );
         $props = [];
         foreach ($props_raw as $prop_raw) {
             $props[$prop_raw->getName()] = $prop_raw;
@@ -237,31 +210,34 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
         return $allowedAttributes;
     }
 
-    /**
-     * @param object $object
-     * @param string $prop_name
-     * @param null $format
-     * @param array $context
-     *
-     * @return mixed
-     */
-    protected function getAttributeValue($object, $prop_name, $format = null, array $context = [])
-    {
+    protected function getAttributeValue(
+        object $object,
+        string $prop_name,
+        string $format = null,
+        array $context = []
+    ): mixed {
         $form_mode = $context[self::NORMALIZE_TO_IDENTIFIERS] ?? false;
 
         if (isset($context[self::CLASS_METADATA]->associationMappings[$prop_name])) {
-            $annotation = $this->annotationReader->getPropertyAnnotation(
-                new ReflectionProperty(get_class($object), $prop_name),
+            $deepNormalizeAttrs = (new ReflectionClass($object))->getProperty($prop_name)->getAttributes(
                 DeepNormalize::class
             );
+            if (!empty($deepNormalizeAttrs)) {
+                /** @var DeepNormalize $deepNormalize */
+                $deepNormalize = current($deepNormalizeAttrs)->newInstance();
 
-            $deep = ($annotation instanceof DeepNormalize)
-                ? $annotation->getDeepNormalize()
-                : false;
+                $deep = $deepNormalize->getDeepNormalize();
+            } else {
+                $deep = false;
+            }
 
             if (!$deep) {
-                throw new NoGetterAvailableException(sprintf('Deep normalization disabled for property %s.',
-                    $prop_name));
+                throw new NoGetterAvailableException(
+                    sprintf(
+                        'Deep normalization disabled for property %s.',
+                        $prop_name
+                    )
+                );
             }
 
             $prop_val = $this->getProperty($object, $prop_name);
@@ -271,21 +247,20 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
                 if (count($prop_val) > 0) {
                     foreach ($prop_val as $val_obj) {
                         if ($form_mode) {
-                            $obj_meta = $this->em->getClassMetadata(get_class($val_obj));
-                            $id_field = $obj_meta->identifier;
+                            $id_field = $this->em->getClassMetadata(get_class($val_obj))->identifier;
 
                             if ($id_field && count($id_field) === 1) {
                                 $return_val[] = $this->getProperty($val_obj, $id_field[0]);
                             }
                         } else {
-                            $return_val[] = $this->serializer->normalize($val_obj, $format, $context);
+                            $return_val[] = $this->normalizer->normalize($val_obj, $format, $context);
                         }
                     }
                 }
                 return $return_val;
             }
 
-            return $this->serializer->normalize($prop_val, $format, $context);
+            return $this->normalizer->normalize($prop_val, $format, $context);
         }
 
         $value = $this->getProperty($object, $prop_name);
@@ -300,9 +275,8 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
      * @param object $entity
      * @param string $key
      *
-     * @return mixed|null
      */
-    protected function getProperty($entity, $key)
+    protected function getProperty(object $entity, string $key): mixed
     {
         // Default to "getStatus", "getConfig", etc...
         $getter_method = $this->getMethodName($key, 'get');
@@ -324,10 +298,8 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
      *
      * @param string $var
      * @param string $prefix
-     *
-     * @return string
      */
-    protected function getMethodName($var, $prefix = ''): string
+    protected function getMethodName(string $var, string $prefix = ''): string
     {
         return $this->inflector->camelize(($prefix ? $prefix . '_' : '') . $var);
     }
@@ -339,8 +311,13 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
      * @param string|null $format
      * @param array $context
      */
-    protected function setAttributeValue($object, $field, $value, $format = null, array $context = []): void
-    {
+    protected function setAttributeValue(
+        object $object,
+        string $field,
+        mixed $value,
+        ?string $format = null,
+        array $context = []
+    ): void {
         if (isset($context[self::ASSOCIATION_MAPPINGS][$field])) {
             // Handle a mapping to another entity.
             $mapping = $context[self::ASSOCIATION_MAPPINGS][$field];
@@ -348,8 +325,12 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
             if ('one' === $mapping['type']) {
                 if (empty($value)) {
                     $this->setProperty($object, $field, null);
-                } elseif (($field_item = $this->em->find($mapping['entity'], $value)) instanceof $mapping['entity']) {
-                    $this->setProperty($object, $field, $field_item);
+                } else {
+                    /** @var class-string $entity */
+                    $entity = $mapping['entity'];
+                    if (($field_item = $this->em->find($entity, $value)) instanceof $entity) {
+                        $this->setProperty($object, $field, $field_item);
+                    }
                 }
             } elseif ($mapping['is_owning_side']) {
                 $collection = $this->getProperty($object, $field);
@@ -359,8 +340,11 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
 
                     if ($value) {
                         foreach ((array)$value as $field_id) {
-                            if (($field_item = $this->em->find($mapping['entity'],
-                                    $field_id)) instanceof $mapping['entity']) {
+                            /** @var class-string $entity */
+                            $entity = $mapping['entity'];
+
+                            $field_item = $this->em->find($entity, $field_id);
+                            if ($field_item instanceof $entity) {
                                 $collection->add($field_item);
                             }
                         }
@@ -377,9 +361,8 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
      * @param string $key
      * @param mixed $value
      *
-     * @return mixed|null
      */
-    protected function setProperty($entity, $key, $value)
+    protected function setProperty(object $entity, string $key, mixed $value): mixed
     {
         $method_name = $this->getMethodName($key, 'set');
 
@@ -388,73 +371,81 @@ class DoctrineEntityNormalizer extends AbstractNormalizer
         }
 
         $method = new ReflectionMethod(get_class($entity), $method_name);
-        $first_param = $method->getParameters()[0];
+        $firstParam = $method->getParameters()[0];
 
-        if ($first_param->hasType()) {
-            /** @var ReflectionNamedType $firstParamTypeObj */
-            $firstParamTypeObj = $first_param->getType();
-            $first_param_type = $firstParamTypeObj->getName();
-
-            switch ($first_param_type) {
-                case 'DateTime':
-                    if (!($value instanceof DateTime)) {
-                        if (!is_numeric($value)) {
-                            $value = strtotime($value . ' UTC');
-                        }
-
-                        $dt = new DateTime;
-                        $dt->setTimestamp($value);
-                        $value = $dt;
-                    }
-                    break;
-
-                case 'int':
-                    if ($value === null) {
-                        if (!$first_param->allowsNull()) {
-                            $value = 0;
-                        }
-                    } else {
-                        $value = (int)$value;
-                    }
-                    break;
-
-                case 'float':
-                    if ($value === null) {
-                        if (!$first_param->allowsNull()) {
-                            $value = 0.0;
-                        }
-                    } else {
-                        $value = (float)$value;
-                    }
-                    break;
-
-                case 'bool':
-                    if ($value === null) {
-                        if (!$first_param->allowsNull()) {
-                            $value = false;
-                        }
-                    } else {
-                        $value = (bool)$value;
-                    }
-                    break;
-            }
-        }
-
+        $value = $this->castValue($firstParam, $value);
         return $entity->$method_name($value);
     }
 
-    /**
-     * @param object|string $class
-     *
-     * @return bool
-     */
-    protected function isEntity($class): bool
+    protected function castValue(ReflectionParameter $firstParam, mixed $value): mixed
+    {
+        if (!$firstParam->hasType()) {
+            return $value;
+        }
+
+        $firstParamTypeObj = $firstParam->getType();
+
+        if (
+            !($firstParamTypeObj instanceof ReflectionNamedType)
+            || !$firstParamTypeObj->isBuiltin()
+        ) {
+            return $value;
+        }
+
+        switch ($firstParamTypeObj->getName()) {
+            case 'string':
+                if ($value === null) {
+                    if (!$firstParam->allowsNull()) {
+                        return '';
+                    }
+                } else {
+                    return (string)$value;
+                }
+                break;
+
+            case 'int':
+                if ($value === null) {
+                    if (!$firstParam->allowsNull()) {
+                        return 0;
+                    }
+                } else {
+                    return (int)$value;
+                }
+                break;
+
+            case 'float':
+                if ($value === null) {
+                    if (!$firstParam->allowsNull()) {
+                        return 0.0;
+                    }
+                } else {
+                    return (float)$value;
+                }
+                break;
+
+            case 'bool':
+                if ($value === null) {
+                    if (!$firstParam->allowsNull()) {
+                        return false;
+                    }
+                } else {
+                    return (bool)$value;
+                }
+                break;
+        }
+
+        return $value;
+    }
+
+    protected function isEntity(mixed $class): bool
     {
         if (is_object($class)) {
             $class = ($class instanceof Proxy || $class instanceof GhostObjectInterface)
                 ? get_parent_class($class)
                 : get_class($class);
-        } elseif (!is_string($class)) {
+        }
+
+        if (!is_string($class)) {
             return false;
         }
 
