@@ -6,12 +6,12 @@ use DateTimeZone;
 
 class Timezone
 {
-    public static function l(DateTime $date_time)
+    public static function l(DateTime $date_time): DateTime
     {
         return self::localize($date_time);
     }
 
-    public static function localize(DateTime $date_time)
+    public static function localize(DateTime $date_time): DateTime
     {
         $tz_name = date_default_timezone_get();
         $tz = new DateTimeZone($tz_name);
@@ -61,8 +61,7 @@ class Timezone
         $utc = new DateTimeZone('UTC');
         $dt = new DateTime('now', $utc);
 
-        $current_tz = new DateTimeZone($tz);
-        $offset = $current_tz->getOffset($dt);
+        $offset = (new DateTimeZone($tz))->getOffset($dt);
 
         return (int)($offset / 60);
     }
@@ -82,7 +81,7 @@ class Timezone
         return 'UTC ' . $sign . str_pad($hour, 2, '0', STR_PAD_LEFT) . ':' . str_pad($minutes, 2, '0');
     }
 
-    public static function fetchSelect()
+    public static function fetchSelect(): array
     {
         static $tz_select;
 
