@@ -1,12 +1,118 @@
 <?php
+
+declare(strict_types=1);
+
 namespace App;
 
+use App\Http\RouterInterface;
+use App\Http\ServerRequest;
+use App\Traits\RequestAwareTrait;
+use Doctrine\Inflector\InflectorFactory;
 use League\Plates\Engine;
 use League\Plates\Template\Data;
+use Psr\EventDispatcher\EventDispatcherInterface;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Symfony\Component\VarDumper\Cloner\VarCloner;
+use Symfony\Component\VarDumper\Dumper\CliDumper;
 
 class View extends Engine
 {
+    use RequestAwareTrait;
+
+    public function __construct(
+        Environment $environment,
+        EventDispatcherInterface $dispatcher,
+        RouterInterface $router,
+        protected Assets $assets
+    ) {
+        parent::__construct($environment->getViewsDirectory(), 'phtml');
+
+        // Add non-request-dependent content.
+        $this->addData(
+            [
+                'environment' => $environment,
+                'router' => $router,
+                'assets' => $this->assets,
+            ]
+        );
+
+        $this->registerFunction(
+            'escapeJs',
+            function ($string) {
+                return json_encode($string, JSON_THROW_ON_ERROR);
+            }
+        );
+
+        $this->registerFunction(
+            'mailto',
+            function ($address, $link_text = null) {
+                $address = substr(chunk_split(bin2hex(" $address"), 2, ';&#x'), 3, -3);
+                $link_text = $link_text ?? $address;
+                return '<a href="mailto:' . $address . '">' . $link_text . '</a>';
+            }
+        );
+
+        $this->registerFunction(
+            'pluralize',
+            function ($word, $num = 0) {
+                if ((int)$num === 1) {
+                    return $word;
+                }
+
+                return InflectorFactory::create()->build()->pluralize($word);
+            }
+        );
+
+        $this->registerFunction(
+            'truncate',
+            function ($text, $length = 80) {
+                return Utilities\Strings::truncateText($text, $length);
+            }
+        );
+
+        $this->registerFunction(
+            'truncateUrl',
+            function ($url) {
+                return Utilities\Strings::truncateUrl($url);
+            }
+        );
+
+        $this->registerFunction(
+            'link',
+            function ($url, $external = true, $truncate = true) {
+                $url = htmlspecialchars($url, ENT_QUOTES);
+
+                $a = ['href="' . $url . '"'];
+                if ($external) {
+                    $a[] = 'target="_blank"';
+                }
+
+                $a_body = ($truncate) ? Utilities\Strings::truncateUrl($url) : $url;
+                return '<a ' . implode(' ', $a) . '>' . $a_body . '</a>';
+            }
+        );
+
+        $dispatcher->dispatch(new Event\BuildView($this));
+    }
+
+    public function setRequest(?ServerRequestInterface $request): void
+    {
+        $this->assets = $this->assets->withRequest($request);
+        $this->request = $request;
+
+        if (null !== $request) {
+            $this->addData(
+                [
+                    'assets' => $this->assets,
+                    'request' => $request,
+                    'router' => $request->getAttribute(ServerRequest::ATTR_ROUTER),
+                    'flash' => $request->getAttribute(ServerRequest::ATTR_SESSION_FLASH),
+                ]
+            );
+        }
+    }
+
     public function reset(): void
     {
         $this->data = new Data();
@@ -15,8 +121,6 @@ class View extends Engine
     /**
      * @param string $name
      * @param array $data
-     *
-     * @return string
      */
     public function fetch(string $name, array $data = []): string
     {
@@ -27,19 +131,19 @@ class View extends Engine
      * Trigger rendering of template and write it directly to the PSR-7 compatible Response object.
      *
      * @param ResponseInterface $response
-     * @param string $template_name
-     * @param array $template_args
-     *
-     * @return ResponseInterface
+     * @param string $templateName
+     * @param array $templateArgs
      */
     public function renderToResponse(
         ResponseInterface $response,
-        $template_name,
-        array $template_args = []
+        string $templateName,
+        array $templateArgs = []
     ): ResponseInterface {
-        $template = $this->render($template_name, $template_args);
+        $template = $this->render($templateName, $templateArgs);
 
         $response->getBody()->write($template);
-        return $response->withHeader('Content-type', 'text/html; charset=utf-8');
+        $response = $response->withHeader('Content-type', 'text/html; charset=utf-8');
+
+        return $this->assets->writeCsp($response);
     }
 }
