@@ -4,11 +4,8 @@ ini_set('display_errors', 1);
 
 $autoloader = require dirname(__DIR__) . '/vendor/autoload.php';
 
-$app = App\AppFactory::create($autoloader, [
-    App\Settings::BASE_DIR => dirname(__DIR__),
-    App\Settings::IS_DOCKER => true,
-    App\Settings::ENABLE_DATABASE => false,
-    App\Settings::ENABLE_REDIS => false,
+$app = App\AppFactory::createApp($autoloader, [
+    App\Environment::BASE_DIR => dirname(__DIR__),
 ]);
 
 $app->run();
